the terrain generator doesnt have those gaps anymore but as i dont have a way of storing and loading curved graphs yet i cant get good terrain gen

+++ game/header/external/fastNoiseLite.h
@@ -53,8 +53,8 @@
 #define FASTNOISELITE_H
 
 // Switch between using floats or doubles for input position
-typedef float FNLfloat;
-//typedef double FNLfloat;
+// typedef float FNLfloat;
+typedef double FNLfloat;
 
 #if defined(__cplusplus)
 extern "C" {
